refactor(core): extract monitoring and WAL/lock bootstrap (#214)
Made-with: Cursor
